refactor(domain): separate domains to different sub-packages
