+++ src/main/java/kitchenpos/application/ProductService.java
@@ -5,7 +5,7 @@ import java.util.stream.Collectors;
 import kitchenpos.application.dto.request.CreateProductDto;
 import kitchenpos.application.dto.response.ProductDto;
 import kitchenpos.dao.ProductDao;
-import kitchenpos.domain.Product;
+import kitchenpos.domain.menu.Product;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
