+++ src/main/java/kitchenpos/domain/menu/ProductQuantity.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.menu;
 
 import java.math.BigDecimal;
