+++ src/main/java/kitchenpos/domain/order/OrderStatus.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.order;
 
 import java.util.Arrays;
 import java.util.List;
