+++ src/main/java/kitchenpos/ui/dto/OrderStatusRequestDto.java
@@ -1,7 +1,7 @@
 package kitchenpos.ui.dto;
 
 import kitchenpos.application.dto.request.UpdateOrderStatusDto;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.domain.order.OrderStatus;
 
 public class OrderStatusRequestDto {
