+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 import java.util.List;
 import kitchenpos.application.dto.request.CreateProductDto;
 import kitchenpos.application.dto.response.ProductDto;
-import kitchenpos.domain.Product;
+import kitchenpos.domain.menu.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
